% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{caseloads}
\alias{caseloads}
\title{Relative Caseloads by Date}
\usage{
caseloads(x, start = "2020-03-01", end = Sys.Date())
}
\arguments{
\item{x}{dat from grabDat}

\item{start}{first day - defaults to March 1st, 2020}

\item{end}{last day - defaults to today}
}
\value{
List of two tibbles: References (RNAME/LENGTH) and Mapping (ID/PN/VN/CL)
}
\description{
Relative Caseloads by Date
}
