% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats2results.R
\name{complete_results}
\alias{complete_results}
\title{Complete evaluation results}
\usage{
complete_results(results, benchmark, alpha, analysis)
}
\arguments{
\item{results}{"cases_results" object, i.e. result of \link{evaluate}}

\item{benchmark}{numeric, vector of benchmark values}

\item{alpha}{numeric, significance level}

\item{analysis}{character, either "co-primary" or "full"}
}
\value{
"cases_results" object
}
\description{
Complete evaluation results
}
\details{
Not exported, but applied at the end of evaluate by default
}
