% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat_ar1}
\alias{cormat_ar1}
\title{Create an AR(1) correlation matrix}
\usage{
cormat_ar1(m, rho, d = TRUE)
}
\arguments{
\item{m}{(numeric) \cr dimensions of the (square) matrix}

\item{rho}{(numeric) \cr correlation parameter in (0,1)}

\item{d}{(logical | numeric) \cr binary vector of length m, whereby TRUE/FALSE (alternatively 1/0)
indicate active/inactive components of underlying random vector.}
}
\value{
(matrix) \cr
AR(1) correlation matrix R with entries \eqn{R_{ij} = \rho^{|i-j|}}
}
\description{
Create an AR(1) correlation matrix
}
