% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data_prb.R
\name{draw_data_prb}
\alias{draw_data_prb}
\title{Sample binary data (single sample)}
\usage{
draw_data_prb(n = 100, pr = c(0.8, 0.8), R = diag(length(pr)))
}
\arguments{
\item{n}{(numeric) \cr integer, sample size}

\item{pr}{(numeric) \cr vector with marginal success probabilities}

\item{R}{(matrix) \cr square correlation matrix}
}
\value{
(matrix) \cr matrix with n rows and length(pr) columns of randomly generated binary (0, 1) data
}
\description{
This function is wrapper for \code{\link[bindata]{rmvbin}}.
}
