% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{equiv}
\alias{equiv}
\title{Equivalent Counts}
\usage{
equiv(x, province)
}
\arguments{
\item{x}{dl$calc}
}
\value{
table of what that number (as a percentage of the population) would look like in other provinces
}
\description{
Sometimes it's useful to know what the direct daily count's equivalent would be in other provinces based on population
Recommended to filter to a smaller date set that is more useful than the overall cases
}
