% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grabDat}
\alias{grabDat}
\title{Get Population and Cases Data}
\usage{
grabDat()
}
\value{
List of three tibbles: pop (population data from statscan), dat (covid cases from canada.ca), calc (combined dat and pop along with relative caseload: calculation of pct cases and diff from expected based on population)
}
\description{
Returns a list of tibbles
}
