% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makelineplot}
\alias{makelineplot}
\title{Line Plots}
\usage{
makelineplot(x, k = 7)
}
\arguments{
\item{x}{dl$calc}

\item{k}{rolling average in days}
}
\value{
dope plot
}
\description{
Create lineplots
}
