% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{the value to be piped in \code{rhs}}

\item{rhs}{the function to be applied with first argument \code{lhs}}
}
\value{
the result of \code{rhs(lhs, ...)}
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
