% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr}
\alias{pr}
\title{Province Names}
\format{
A 14 element character vector
\describe{
\item{}{"Canada", "Alberta", "British Columbia", "Manitoba", "New Brunswick", "Newfoundland and Labrador", "Northwest Territories", "Nova Scotia", "Nunavut", "Ontario", "Prince Edward Island", "Quebec", "Saskatchewan", "Yukon"}
...
}
}
\usage{
pr
}
\description{
Including Canada
}
\keyword{datasets}
