% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provs}
\alias{provs}
\title{Provinces in order}
\format{
A 13 element ordered list
\describe{
\item{}{"Canada", "Alberta", "British Columbia", "Manitoba", "New Brunswick", "Newfoundland and Labrador", "Northwest Territories", "Nova Scotia", "Nunavut", "Ontario", "Prince Edward Island", "Quebec", "Saskatchewan", "Yukon"}
...
}
}
\usage{
provs
}
\description{
Provinces in order
}
\keyword{datasets}
