% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize evaluation results}
\usage{
visualize(x, ...)
}
\arguments{
\item{x, }{a \code{cases_results} object, see \code{\link{evaluate}}}

\item{...}{further arguments (currently ignored)}
}
\value{
a ggplot
}
\description{
Visualize evaluation results
}
\details{
+++ early development version (only alternative = "greater" is supported) +++
}
