% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\docType{package}
\name{cases}
\alias{cases}
\title{\code{cases} package}
\description{
Evaluate the accuracy of multiple (candidate) classifiers in several subgroups
}
\details{
See the README on
\href{https://github.com/maxwestphal/cases}{GitHub}
}
