% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\docType{package}
\name{cases}
\alias{cases}
\title{\code{cases} package}
\description{
Enables simultaneous statistical inference for the accuracy of multiple classifiers in multiple subgroups (strata). For instance, allows to perform multiple comparisons in diagnostic accuracy studies with co-primary endpoints sensitivity and specificity. (Westphal, Max, and Antonia Zapf. "Statistical Inference for Diagnostic Test Accuracy Studies with Multiple Comparisons." arXiv:2105.13469 (2021).)
}
\details{
See the vignettes
vignette()
}
