% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat_equi}
\alias{cormat_equi}
\title{Create an equicorrelation matrix}
\usage{
cormat_equi(m, rho, d = TRUE)
}
\arguments{
\item{m}{integer, dimension}

\item{rho}{numeric, correlation parameter in (0,1)}

\item{d}{binary vector of length m, whereby TRUE/FALSE (alternatively 1/0)
indicate active/inactive components of underlying random vector.}
}
\value{
\eqn{R_{ij} = \rho, i\neq j}
}
\description{
Create an equicorrelation matrix
}
