% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize evaluation results}
\usage{
visualize(x, ...)
}
\arguments{
\item{x}{\code{cases_results} \cr produced by \code{\link{evaluate}}}

\item{...}{\code{any} \cr further arguments (currently ignored)}
}
\value{
a ggplot
}
\description{
\strong{Currently, this implementation is only intended for situations with ...}
\itemize{
\item two groups (e.g. healthy (<-> specificity) and diseased (<-> sensitivity))
\item alternative = "greater"
\item contrast = define_contrast("raw)
}
}
